🖌️ Modify : 마커 클릭시 기존의 커스텀 오버레이 삭제 및 클릭한 오버레이 렌더링
+ 마커 클릭시 커스텀 오버레이 전환 구현

+ debounceFunc 모듈화

CO-353
