+++ src/utils/index.ts
@@ -14,3 +14,4 @@ export * from "./formData"
 export * from "./dragAndDrop"
 export * from "./fileToObjectUrl"
 export * from "./objectUrlToFile"
+export * from "./debounceFunc"
